+++ examples/openai-chat/src/main.rs
@@ -2,7 +2,7 @@ use opai::{
     chats::message::{CreateChatCompletion, Message, Role},
     client::Client,
     config::Config,
-    models::gpt::GptModel,
+    models::{gpt::Gpt, Model},
 };
 
 #[tokio::main]
@@ -15,7 +15,7 @@ async fn main() {
         role: Role::User,
         name: None,
     }];
-    let request = CreateChatCompletion::new(GptModel::GPT4o, messages);
+    let request = CreateChatCompletion::new(Model::Gpt(Gpt::GPT4), messages);
     let completion = client.chat.create_completion(request).await.unwrap();
     println!("{:?}", completion);
 }